input done and working
