--- hello.txt
@@ -1,1 +1,0 @@
-was geht
